+++ cmd/argoexec/commands/wait.go
@@ -4,7 +4,6 @@ import (
 	"time"
 
 	"github.com/argoproj/argo/util/stats"
-	"github.com/argoproj/argo/workflow/common"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -21,36 +20,36 @@ var waitCmd = &cobra.Command{
 
 func waitContainer(cmd *cobra.Command, args []string) {
 	wfExecutor := initExecutor()
-	defer wfExecutor.AnnotatePanic()
+	defer wfExecutor.HandleError()
 	defer stats.LogStats()
 	stats.StartStatsTicker(5 * time.Minute)
 
 	// Wait for main container to complete and kill sidecars
 	err := wfExecutor.Wait()
 	if err != nil {
-		_ = wfExecutor.AddAnnotation(common.AnnotationKeyNodeMessage, err.Error())
-		log.Errorf("Error waiting on main container to be ready, %+v", err)
+		wfExecutor.AddError(err)
+		log.Errorf("Error on wait, %+v", err)
 	}
 	err = wfExecutor.SaveArtifacts()
 	if err != nil {
-		_ = wfExecutor.AddAnnotation(common.AnnotationKeyNodeMessage, err.Error())
+		wfExecutor.AddError(err)
 		log.Fatalf("Error saving output artifacts, %+v", err)
 	}
 	// Saving output parameters
 	err = wfExecutor.SaveParameters()
 	if err != nil {
-		_ = wfExecutor.AddAnnotation(common.AnnotationKeyNodeMessage, err.Error())
+		wfExecutor.AddError(err)
 		log.Fatalf("Error saving output parameters, %+v", err)
 	}
 	// Capture output script result
 	err = wfExecutor.CaptureScriptResult()
 	if err != nil {
-		_ = wfExecutor.AddAnnotation(common.AnnotationKeyNodeMessage, err.Error())
+		wfExecutor.AddError(err)
 		log.Fatalf("Error capturing script output, %+v", err)
 	}
 	err = wfExecutor.AnnotateOutputs()
 	if err != nil {
-		_ = wfExecutor.AddAnnotation(common.AnnotationKeyNodeMessage, err.Error())
+		wfExecutor.AddError(err)
 		log.Fatalf("Error annotating outputs, %+v", err)
 	}
 }